backend/requirements.txt
psycopg2-binary==2.9.5
blacksheep==1.2.8
python-dotenv==0.21.0
uvicorn==0.20.0
fastapi
cchardet